[WFLY-19657] Update AppClient parsing to use the "Schemas" approach with stability level support.
